refactor(upload): break the upload helper to god-file import cycle
The 2026-06 extraction (#5574/#5589/#5601) moved PendingUploadStore,
UploadRetryPolicy and UploadProgressReporter out of upload_manager.dart,
but left a back-edge: both UploadRetryPolicy and UploadProgressReporter
still imported upload_manager.dart, because four symbols they need were
declared inside the god file. The helpers were therefore not actually
independent of it, and the cycle would block any future package lift.

Move the four shared symbols to where they belong:

  UploadRetryConfig, videoProgressShare -> upload/upload_config.dart (new)
  UploadMetrics                         -> upload/upload_metrics.dart (new)
  BlossomUploadFailureException         -> upload/upload_session_errors.dart

The exception joins the existing shared-errors file, which already owns
isExpiredResumableSessionError and already imports blossom_upload_service.

upload_manager.dart re-exports all four, so the ~11 test files that reach
them through it are untouched. videoProgressShare was a static on
UploadManager and is now a top-level const, so its three test call sites
lose the class qualifier.

Also drops three stale `REFACTORED: Removed ChangeNotifier` comments and
one dangling `/// Upload target options` doc comment that documented
nothing.

No behaviour change. Refs #6935
